refactor: removed invocation provider.
also moved handler tests into its own file.
