+++ zeebe/exporters/camunda-exporter/src/test/java/io/camunda/exporter/CamundaExporterIT.java
@@ -17,7 +17,6 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.CALLS_REAL_METHODS;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.times;
@@ -27,58 +26,28 @@ import static org.mockito.Mockito.when;
 import io.camunda.exporter.cache.ProcessCacheLoaderFactory;
 import io.camunda.exporter.config.ConnectionTypes;
 import io.camunda.exporter.config.ExporterConfiguration;
-import io.camunda.exporter.handlers.DecisionEvaluationHandler;
-import io.camunda.exporter.handlers.EventFromJobHandler;
-import io.camunda.exporter.handlers.ExportHandler;
-import io.camunda.exporter.handlers.ListViewFlowNodeFromJobHandler;
-import io.camunda.exporter.handlers.ListViewFlowNodeFromProcessInstanceHandler;
-import io.camunda.exporter.handlers.ListViewProcessInstanceFromProcessInstanceHandler;
-import io.camunda.exporter.handlers.MetricFromProcessInstanceHandler;
-import io.camunda.exporter.handlers.UserTaskProcessInstanceHandler;
 import io.camunda.exporter.schema.SchemaTestUtil;
-import io.camunda.exporter.utils.CamundaExporterHandlerITInvocationProvider;
 import io.camunda.exporter.utils.CamundaExporterITInvocationProvider;
 import io.camunda.exporter.utils.SearchClientAdapter;
 import io.camunda.webapps.schema.descriptors.IndexDescriptor;
 import io.camunda.webapps.schema.descriptors.IndexTemplateDescriptor;
-import io.camunda.webapps.schema.entities.ExporterEntity;
 import io.camunda.zeebe.exporter.api.ExporterException;
 import io.camunda.zeebe.exporter.api.context.Context;
 import io.camunda.zeebe.exporter.test.ExporterTestConfiguration;
 import io.camunda.zeebe.exporter.test.ExporterTestContext;
 import io.camunda.zeebe.exporter.test.ExporterTestController;
 import io.camunda.zeebe.protocol.record.Record;
-import io.camunda.zeebe.protocol.record.RecordValue;
 import io.camunda.zeebe.protocol.record.ValueType;
-import io.camunda.zeebe.protocol.record.intent.JobIntent;
-import io.camunda.zeebe.protocol.record.value.BpmnElementType;
-import io.camunda.zeebe.protocol.record.value.DecisionEvaluationRecordValue;
-import io.camunda.zeebe.protocol.record.value.EvaluatedDecisionValue;
-import io.camunda.zeebe.protocol.record.value.ImmutableDecisionEvaluationRecordValue;
-import io.camunda.zeebe.protocol.record.value.ImmutableEvaluatedDecisionValue;
-import io.camunda.zeebe.protocol.record.value.ImmutableJobRecordValue;
-import io.camunda.zeebe.protocol.record.value.ImmutableProcessInstanceRecordValue;
-import io.camunda.zeebe.protocol.record.value.JobRecordValue;
-import io.camunda.zeebe.protocol.record.value.ProcessInstanceRecordValue;
 import io.camunda.zeebe.protocol.record.value.UserRecordValue;
 import io.camunda.zeebe.test.broker.protocol.ProtocolFactory;
 import io.camunda.zeebe.test.util.testcontainers.TestSearchContainers;
-import java.io.IOException;
 import java.time.Duration;
-import java.time.OffsetDateTime;
-import java.time.temporal.ChronoUnit;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
-import java.util.function.Function;
-import java.util.function.Supplier;
 import java.util.stream.Stream;
 import org.awaitility.Awaitility;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.TestInstance.Lifecycle;
 import org.junit.jupiter.api.TestTemplate;
@@ -95,7 +64,6 @@ import org.testcontainers.containers.GenericContainer;
  */
 @TestInstance(Lifecycle.PER_CLASS)
 @ExtendWith(CamundaExporterITInvocationProvider.class)
-@Disabled("flaky test")
 final class CamundaExporterIT {
 
   private final ProtocolFactory factory = new ProtocolFactory();
@@ -338,145 +306,5 @@ final class CamundaExporterIT {
     when(provider.getIndexTemplateDescriptors()).thenReturn(templateDescriptors);
 
     return provider;
-  }
-
-  @Nested
-  @ExtendWith(CamundaExporterHandlerITInvocationProvider.class)
-  class ExportHandlerTests {
-
-    final Map<Class<?>, Function<ExportHandler<?, ?>, Record<?>>> customRecordGenerators =
-        new HashMap<>(
-            Map.of(
-                EventFromJobHandler.class,
-                this::jobRecordGenerator,
-                ListViewFlowNodeFromJobHandler.class,
-                this::jobRecordGenerator,
-                ListViewProcessInstanceFromProcessInstanceHandler.class,
-                (handler) -> processInstanceRecordGenerator(handler, BpmnElementType.PROCESS),
-                ListViewFlowNodeFromProcessInstanceHandler.class,
-                (handler) ->
-                    processInstanceRecordGenerator(handler, BpmnElementType.BOUNDARY_EVENT),
-                DecisionEvaluationHandler.class,
-                this::decisionEvalRecordGenerator,
-                MetricFromProcessInstanceHandler.class,
-                (handler) -> processInstanceRecordGenerator(handler, BpmnElementType.PROCESS),
-                UserTaskProcessInstanceHandler.class,
-                (handler) -> processInstanceRecordGenerator(handler, BpmnElementType.PROCESS)));
-
-    @TestTemplate
-    @SuppressWarnings("unchecked")
-    <S extends ExporterEntity<S>, T extends RecordValue> void allHandlerTestsWithInvocationProvider(
-        final CamundaExporter exporter,
-        final SearchClientAdapter clientAdapter,
-        final ExportHandler<S, T> handler)
-        throws IOException {
-
-      final var currentTime = OffsetDateTime.now().truncatedTo(ChronoUnit.MILLIS);
-      try (final var mocked = mockStatic(OffsetDateTime.class)) {
-        mocked.when(OffsetDateTime::now).thenReturn(currentTime);
-
-        final var record =
-            (Record<T>)
-                customRecordGenerators
-                    .getOrDefault(handler.getClass(), this::defaultRecordGenerator)
-                    .apply(handler);
-        customRecordGenerators.put(handler.getClass(), (ignored) -> record);
-
-        final var entityId = handler.generateIds(record).getFirst();
-        final var expectedEntity = handler.createNewEntity(entityId);
-        handler.updateEntity(record, expectedEntity);
-        exporter.export(record);
-
-        final var responseEntity =
-            clientAdapter.get(
-                expectedEntity.getId(), handler.getIndexName(), handler.getEntityType());
-
-        assertThat(responseEntity)
-            .describedAs(
-                "Handler [%s] correctly handles a [%s] record",
-                handler.getClass().getSimpleName(), handler.getHandledValueType())
-            .isEqualTo(expectedEntity);
-      }
-    }
-
-    private <S extends ExporterEntity<S>, T extends RecordValue> Record<T> recordGenerator(
-        final ExportHandler<S, T> handler, final Supplier<Record<T>> createRecord) {
-      // Sometimes the factory generates record with intents that are not supported by the handler.
-      var record = createRecord.get();
-      do {
-        record = createRecord.get();
-      } while (!handler.handlesRecord(record));
-      return record;
-    }
-
-    private <S extends ExporterEntity<S>, T extends RecordValue> Record<T> jobRecordGenerator(
-        final ExportHandler<S, T> handler) {
-      return recordGenerator(
-          handler,
-          () -> {
-            final var jobRecordValue =
-                ImmutableJobRecordValue.builder()
-                    .from(factory.generateObject(JobRecordValue.class))
-                    .withDeadline(System.currentTimeMillis() + 10000)
-                    .build();
-            return factory.generateRecord(
-                ValueType.JOB,
-                r ->
-                    r.withValue((T) jobRecordValue)
-                        .withTimestamp(System.currentTimeMillis())
-                        .withIntent(JobIntent.CREATED));
-          });
-    }
-
-    private <S extends ExporterEntity<S>, T extends RecordValue>
-        Record<T> decisionEvalRecordGenerator(final ExportHandler<S, T> handler) {
-      return recordGenerator(
-          handler,
-          () -> {
-            final ImmutableEvaluatedDecisionValue decisionValue =
-                ImmutableEvaluatedDecisionValue.builder()
-                    .from(factory.generateObject(EvaluatedDecisionValue.class))
-                    .build();
-            final DecisionEvaluationRecordValue decisionRecordValue =
-                ImmutableDecisionEvaluationRecordValue.builder()
-                    .from(factory.generateObject(DecisionEvaluationRecordValue.class))
-                    .withEvaluatedDecisions(List.of(decisionValue))
-                    .build();
-            return factory.generateRecord(
-                ValueType.DECISION_EVALUATION,
-                r ->
-                    r.withValue((T) decisionRecordValue).withTimestamp(System.currentTimeMillis()));
-          });
-    }
-
-    private <S extends ExporterEntity<S>, T extends RecordValue> Record<T> defaultRecordGenerator(
-        final ExportHandler<S, T> handler) {
-      return recordGenerator(
-          handler,
-          () ->
-              factory.generateRecord(
-                  handler.getHandledValueType(), r -> r.withTimestamp(System.currentTimeMillis())));
-    }
-
-    private <S extends ExporterEntity<S>, T extends RecordValue>
-        Record<T> processInstanceRecordGenerator(
-            final ExportHandler<S, T> handler, final BpmnElementType elementType) {
-      return recordGenerator(
-          handler,
-          () -> {
-            final ProcessInstanceRecordValue processInstanceRecordValue =
-                ImmutableProcessInstanceRecordValue.builder()
-                    .from(factory.generateObject(ProcessInstanceRecordValue.class))
-                    .withParentProcessInstanceKey(-1L)
-                    .withBpmnElementType(elementType)
-                    .build();
-            return factory.generateRecord(
-                ValueType.PROCESS_INSTANCE,
-                r ->
-                    r.withIntent(ELEMENT_ACTIVATING)
-                        .withValue((T) processInstanceRecordValue)
-                        .withTimestamp(System.currentTimeMillis()));
-          });
-    }
   }
 }
